add play with LAN to history

+++ CSharp_CaroGame/Form_CoCaRo.cs
@@ -124,6 +124,8 @@ namespace CSharp_CaroGame
                 panel_banco.Enabled = false;
                 Listen();
                 MessageBox.Show("Kết nối thành công. Trận đấu sẵn sàng");
+                socket.Send(new SocketData((int)SocketCommand.MESSNAME, textBox_username.Text, new Point()));
+
             }
             timer1.Stop();
             pgb_Time.Value = 0;
@@ -164,7 +166,7 @@ namespace CSharp_CaroGame
                             return;
                         }
                     }
-                    
+
                 }
 
                 else if (Control.CheDoChoi == 3)
@@ -207,12 +209,20 @@ namespace CSharp_CaroGame
 
         private void ProcessData(SocketData data)
         {
+
             switch (data.Command)
             {
+                case (int)SocketCommand.MESSNAME:
+                    this.Invoke((MethodInvoker)(() =>
+                    {
+                        Control.username1 = data.Message;
 
+                    }));
+                    break;
                 case (int)SocketCommand.NEW_GAME:
                     this.Invoke((MethodInvoker)(() =>
                     {
+
                         NewGame();
                         panel_banco.Enabled = false;
                     }));
@@ -235,6 +245,7 @@ namespace CSharp_CaroGame
 
                     break;
                 case (int)SocketCommand.END_GAME:
+
                     break;
                 case (int)SocketCommand.UNDO:
                     this.Invoke((MethodInvoker)(() =>
@@ -271,7 +282,7 @@ namespace CSharp_CaroGame
                 }
             }
         }
-        
+
         private void btn_Replay_Click(object sender, EventArgs e)
         {
 
@@ -300,7 +311,7 @@ namespace CSharp_CaroGame
             }
         }
 
-        
+
         #endregion
 
         #region Undo, Redo
@@ -362,7 +373,14 @@ namespace CSharp_CaroGame
                 StringSplitOptions.None
             );
                 string filePath = @"D:\\history.txt";
-                System.IO.File.WriteAllLines(filePath, lines);
+                if (!System.IO.File.Exists(filePath))
+                {
+                    System.IO.File.WriteAllLines(filePath, lines);
+                }
+                else
+                {
+                    System.IO.File.WriteAllLines(filePath, lines);
+                }
                 String[] line1 = File.ReadLines(filePath).Take(11).ToArray();
                 System.IO.File.WriteAllLines(filePath, line1);
                 Label[] lb = new Label[11];
@@ -371,25 +389,25 @@ namespace CSharp_CaroGame
                 {
                     String ss = line1[i].ToString();
 
-                    if (ss.Contains("thua"))
+                    if (ss.Contains("thua") || ss.Contains("lose"))
                     {
                         lb[i] = new Label();
                         lb[i].BackColor = Color.Red;
                         lb[i].Height = 23;
                         lb[i].Width = 530;
-                        lb[i].Text = lines[i];
+                        lb[i].Text = line1[i];
                         lb[i].Location = new Point(30, 25 + (i * 30));
                         hs.panel_history.Controls.Add(lb[i]);
                         reader.Close();
 
                     }
-                    if (ss.Contains("thắng"))
+                    if (ss.Contains("thắng") || ss.Contains("win"))
                     {
                         lb[i] = new Label();
                         lb[i].BackColor = Color.Green;
                         lb[i].Height = 23;
                         lb[i].Width = 530;
-                        lb[i].Text = lines[i];
+                        lb[i].Text = line1[i];
                         lb[i].Location = new Point(30, 25 + (i * 30));
                         hs.panel_history.Controls.Add(lb[i]);
                         reader.Close();
@@ -427,6 +445,6 @@ namespace CSharp_CaroGame
             grap.Clear(panel_banco.BackColor);
             timer1.Start();
             pgb_Time.Value = 0;
-        } 
+        }
     }
 }
